requirements.txt
# pip install -r requirements.txt
albumentations==1.1.0
numpy==1.22.1
opencv-python==4.5.5.62
pytest==7.0.0
PyYAML==6.0
segmentation-models-pytorch==0.2.1
tensorboard==2.8.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
